TSE-21 Rename MainRightSide to MainLeftSide
