+++ src/components/pages/order/Main/MainLeftSide/Admin/tabsConfig.tsx
