+++ src/components/pages/order/Main/MainLeftSide/Menu/RibbonAnimated.tsx
